# openapi.yaml
openapi: 3.0.3
info:
  title: Corgi Recommender Service API
  description: |
    API for the Corgi Recommender Service, a personalized recommendations system
    for Mastodon posts. The API handles user interactions, preferences, and generates
    personalized timelines.
    
    ## Key Features
    - **User Interaction Tracking**: Records favorites, bookmarks, and explicit feedback
    - **Personalized Recommendations**: Generates tailored post suggestions based on user activity
    - **Privacy Controls**: Provides user-configurable privacy settings
    - **Mastodon Compatibility**: Returns data in Mastodon-compatible format for easy integration
    - **Transparent Proxy**: Proxies requests to Mastodon instances with added recommendations
  version: 1.0.0
  contact:
    name: Corgi Service Team
    email: support@corgi-recommender.example.com
servers:
  - url: http://localhost:5001
    description: Local development server
  - url: https://api.corgi-recommender.example.com
    description: Production server
...
paths:
  /api/v1/interactions:
    post:
      summary: Log a user interaction with a post
      description: |
        Records user actions like favorites, bookmarks, or post feedback.
        This data is used to improve future recommendations.
      operationId: logInteraction
      tags:
        - interactions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InteractionRequest'
            example:
              user_alias: "user_12345"
              post_id: "post_67890"
              action_type: "favorite"
              context:
                source: "timeline_home"
      responses:
        '200':
          description: Interaction logged successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  interaction_id:
                    type: string
                    description: The ID of the logged interaction
                    example: "interaction_12345"
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/interactions/{post_id}:
    get:
      summary: Get interactions for a specific post
      description: Retrieves all recorded interactions for a single post
      operationId: getPostInteractions
      tags:
        - interactions
      parameters:
        - name: post_id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the post to retrieve interactions for
      responses:
        '200':
          description: Interactions for the post
          content:
            application/json:
              schema:
                type: object
                properties:
                  post_id:
                    type: string
                  interaction_counts:
                    $ref: '#/components/schemas/InteractionCounts'
                  interactions:
                    type: array
                    items:
                      type: object
                      properties:
                        action_type:
                          type: string
                        count:
                          type: integer
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/interactions/counts/batch:
    post:
      summary: Get interaction counts for multiple posts in a single request
      description: Efficiently retrieves interaction counts for a batch of posts
      operationId: batchGetInteractionCounts
      tags:
        - interactions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - post_ids
              properties:
                post_ids:
                  type: array
                  items:
                    type: string
                  description: Array of post IDs to get counts for
            example:
              post_ids: ["post_12345", "post_67890", "post_24680"]
      responses:
        '200':
          description: Interaction counts for the requested posts
          content:
            application/json:
              schema:
                type: object
                properties:
                  counts:
                    type: object
                    additionalProperties:
                      $ref: '#/components/schemas/InteractionCounts'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/interactions/user/{user_id}:
    get:
      summary: Get all interactions for a specific user
      description: Retrieves interactions performed by a single user
      operationId: getUserInteractions
      tags:
        - interactions
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the user to retrieve interactions for
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
          description: Maximum number of interactions to return
      responses:
        '200':
          description: User interactions
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: string
                  interactions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Interaction'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/interactions/favourites:
    get:
      summary: Get all posts favorited by a specific user
      description: Retrieves posts that a user has marked as favorites
      operationId: getUserFavorites
      tags:
        - interactions
      parameters:
        - name: user_id
          in: query
          required: true
          schema:
            type: string
          description: The ID of the user to retrieve favorites for
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
          description: Maximum number of favorites to return
      responses:
        '200':
          description: Favorited posts
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: string
                  favorites:
                    type: array
                    items:
                      $ref: '#/components/schemas/MastodonPost'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  schemas:
    InteractionRequest:
      type: object
      required:
        - post_id
        - action_type
      properties:
        user_alias:
          type: string
          description: Pseudonymized identifier for the user
        user_id:
          type: string
          description: Alternative direct user identifier
        post_id:
          type: string
          description: Unique identifier for the post
        action_type:
          type: string
          description: The type of interaction
          enum:
            - favorite
            - bookmark
            - reblog
            - more_like_this
            - less_like_this
        context:
          type: object
          description: Additional context about the interaction
          properties:
            source:
              type: string
              description: Source of the interaction (e.g., timeline_home)
    Interaction:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the interaction
        user_alias:
          type: string
          description: Pseudonymized identifier for the user
        post_id:
          type: string
          description: Identifier for the post interacted with
        action_type:
          type: string
          description: Type of interaction (favorite, bookmark, etc.)
        created_at:
          type: string
          format: date-time
          description: When the interaction occurred
        context:
          type: object
          description: Additional context about the interaction
          additionalProperties: true
    InteractionCounts:
      type: object
      properties:
        favorites:
          type: integer
          default: 0
        reblogs:
          type: integer
          default: 0
        replies:
          type: integer
          default: 0
        bookmarks:
          type: integer
          default: 0
...
  responses:
    BadRequest:
      description: Bad request due to invalid parameters or payload
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              received:
                type: object
                additionalProperties: true
    NotFound:
      description: The requested resource was not found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              type:
                type: string
              request_id:
                type: string
...